reader writer

+++ pkg/codec/codec.go
@@ -15,7 +15,7 @@ func Decode(b []byte) uint16 {
 
 // [ 0 0 ] first two bytes amount of next bytes info uint16 65535
 
-func WriteHeader(b []byte, req uint16) {
+func WriteHeader(b []byte, req int) {
 	b[0] = byte(req >> 8)
 	b[1] = byte(req)
 }
